Bump kind version
There have been a lot of changes in 0.6 release. One point to note is that even though this is a minor release, there is a breaking change as described in the release notes - https://github.com/kubernetes-sigs/kind/releases. The breaking change is easy to fix in pipeline though - users will just have to change handling the KUBECONFIG as described in the release notes.