fix(win): use correct case sensitive file name
File name was submitted with `CLI` in `AzureDevSpacesCLI` capitalized,
but the script is invoked as a provisioner using the lower-case `Cli`
form as seen in `AzureDevSpacesCli`.

Renaming the file to use the lower-case form to match other provisioning
scripts used in the Windows builds.

Issue introduced in commit 932d0f36cdabe7e40fc8abb46469dcee1922b252