refactor(windows): consistent file naming
Use the same file naming scheme for Windows as is used for Ubuntu, where
the file name includes the Operating System name and release, but
not the target platform.

This also encourages downstream consumers to re-use the same files with
alternative targets by extending the `builders` section and using the
`-only` command line option with Packer to specify the build target.

Closes #1636