FIX Writing output. Incorrect call to variables.
Error calling variables into last step of Azure VM creation. 

ResourceGroupName: The term 'ResourceGroupName' is not recognized as a name of a cmdlet, function, script file, or executable program.
Check the spelling of the name, or if a path was included, verify that the path is correct and try again.