Update MinGW on Windows 2022 (#5729)
* Update MinGW on Windows 2022

Previously, the GitHub virtual environments only had MinGW v8.1.0
installed. That version was released in October of 2018 and, since then,
MinGW has released versions all the way up to v11.2.0. This change
updates the MinGW version to the latest available on Chocolatey for the
Windows 2022 environment. Windows 2019 retains the current v8.1.0
version. Fixes #5530.

* Update MinGW documentation

* Additionally update win-2016 version

* Revert "Update MinGW documentation"

This reverts commit cb7a72ff0f33f8a5bbda15be677b348a05f9fee0.

* fix: modify the pattern used for retrieving the MinGW version