Ensure VS MEF is also initialized in the VS experimental
MEF initialization is very costly and can take up to 15' on first run of VS.

MEF is only initialized when some action in the IDE requires resolving a MEF 
export in VS. The heaviest consumer of MEF in VS is the core editor itself, and 
the existing initialization code was not causing the editor to open at all, leaving 
MEF uninitialized. 

By opening even a (mostly) empty C# file, we cause most key components in 
MEF to be properly initialized once: editor and project system.