Download two prior CodeQL bundles
Previously, our runner image only included the most recent release of CodeQL. In order to improve stability of rollout of new CodeQL versions, we are now downloading the most recent two releases.

The change also updates the Linux script variable names to match that of the Mac script, because they are mostly functionally identical.